Compatibility fixes
Ditch most of the const's for vars and use a function within the loop
to close over the testFn definitions. This is so entryArgs is not
shared among the different iterations of the loop (so async testFn's
end up all using the same entryArgs). It also avoids using the
potentially-incompatible let statement.

Get rid of destructuring.

Version bump.