Go with "describe" functionality
Change the tabletests function into describeTests.  This puts together
the functionality of table tests and describe. Each describeTests is a
describe, titled with the first argument. The second argument is now
the template for the entries.

This includes support for focusing and skipping of describeTests. Test
reformatting and minor tweaks to names.

Do a bit of refactoring to cut down the complexity.